pcal-mode: improve code readability by shadowing labels
PlusCal labels mostly a necessity rather than part of the algorithm.
They simply point places that can't execute atomically, which results
in having them on almost every line. Labels are a hint to TLC much
less for people reading the code. So let's have a separate face for
labels that kind of shadows their visibility. There's a `shadow` face
from which we inherit for that purpose.